Problem 1181005
An instanceName without keyBindings means either a singleton
or a class specification is involved. Both cases are not
supported for the moment by sfcBroker for the associators,
associatorNames, references and referenceNames operations.
Both cases are very rare in CIM processing, when deemed
absolutely needed, please generate a formal request.

cimXmlRequest.c is updated to generate a
CMPI_RC_ERR_NOT_SUPPORTED rc accompanied with the followiing
message: "<operation> operation for classes not supported".