Match startLogging() and closeLogging() calls
The startLogging and closeLogging calls where unmatched. The former
was called during CIMCEnv creation, the latter during client release.

In case of client connect error, this led to a dangling startLogging
since there was no client to release.

This patch moves the closeLogging() from the client to the CIMCEnv
release.

Signed-off-by: Klaus Kämpf <kkaempf@suse.de>