Configurable local socket group permission for daemon clients
(Chris_Poblete@Dell.com)

In our environment where we use Sfcb, services are run as root. One of
the services is a web service that runs worker threads as daemon user.
A worker thread connects to Sfcb through the local socket interface in
Sfcc. The connection is failing since the local socket file permission
is set to 755 for user:group root:root and the worker thread is
running as daemon:daemon.

Solution:

The proposal allows the socket file group permission to be set to a
specific group. The group name is defined in the Sfcb configuration
file. Once set, the group shall have write permission to the socket
file which will allow the local socket connection.  By default, the
group name is not set and the default permission is used.  Security is
preserved since the configuration file is writable only by user root
(when properly installed).  The installation shall be responsible for
assigning a trusted group as applicable.