Fix large https requests: Check SSL buffer before select() on file descriptor
There is a 5000 char buffer for http/https socket reads and larger
requests (http content length) need another read() to get the
remaining data.
Doing this via select() and read() on the socket is fine for http but
fails for https since SSL has its own decrypt/buffer mode.
This made large https requests fail as the select() on the socket
timed out since the data was already in the SSL buffer.

Fix: Use SSL_pending() to check if there is data left in the SSL
buffer before doing a select() on the socket.